:recycle:  style, types 파일로 분리
