+++ src/components/atoms/Upload/style.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+const UploadContainer = styled.div`
+  display: inline-block;
+  cursor: pointer;
+`;
+const UploadInput = styled.input`
+  display: none;
+`;
+
+export { UploadContainer, UploadInput };
