+++ src/components/atoms/Upload/types.ts
@@ -1,0 +1,10 @@
+import { ChangeEvent, ReactNode } from 'react';
+
+export interface UploadProps {
+  onChange(e: ChangeEvent<HTMLInputElement>): void;
+  children?: ReactNode;
+  name?: string;
+  accept?: string;
+  value?: File;
+  dropable: boolean;
+}
